+++ packages/create-kompo/CHANGELOG.md
@@ -1,5 +1,9 @@
 # create-kompo
 
+## 0.1.3-beta.3
+
+## 0.1.3-beta.2
+
 ## 0.1.3-beta.1
 
 ## 0.1.3-beta.0